chore(sdk): split out wasm integ test from example
