+++ crates/sdk-core/tests/fixtures/wasm_task_failure/.gitignore
@@ -1,0 +1,1 @@
+/target/
